Fixed errors in example 9, moved example 10 to Vite
