+++ example-10-routing-02/src/GalleryPage.jsx
