--- example-10-routing-02/src/NavBar.js
@@ -1,11 +1,0 @@
-import styles from './NavBar.module.css';
-import { NavLink } from 'react-router-dom';
-
-export default function NavBar() {
-    return (
-        <div className={styles.navBar}>
-            <NavLink to="/articles" className={({ isActive }) => isActive && styles.activeLink}>Articles</NavLink>
-            <NavLink to="/gallery" className={({ isActive }) => isActive && styles.activeLink}>Gallery</NavLink>
-        </div>
-    );
-}
